# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-Dashboard API description.",
        "title": "Swagger E-Dashboard API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/edashboard",
    "paths": {
        "/admission/report-popup": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admission"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "sDate",
                        "name": "sDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dateType",
                        "name": "dateType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admission/report/day": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admission"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "days",
                        "name": "days",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sHospitalCode",
                        "name": "sHospitalCode",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admission/report/month": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admission"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_user_session_id",
                        "name": "vx_user_session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_user_name",
                        "name": "vx_user_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "months",
                        "name": "months",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sHospitalCode",
                        "name": "sHospitalCode",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/appointment/statistics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/appointment/statistics-popup": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "sDate",
                        "name": "sDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sType",
                        "name": "sType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_comp_name",
                        "name": "vx_comp_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "vx_environment",
                        "name": "vx_environment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}